Added chime.wav
Imported chime.wav in App.js, added a playBeep() method to handle this
in the code. The chime plays inside the countdown method when the
minutes and seconds are both strictly equal to zero.

Still haven't added the other requirements for the sound portion of
this, but that seems easy enough.